Extract to separate file - private API
This will be removed soon.
