+++ ruby_event_store/lib/ruby_event_store.rb
@@ -16,6 +16,7 @@ require 'ruby_event_store/expected_version'
 require 'ruby_event_store/serialized_record'
 require 'ruby_event_store/transform_keys'
 require 'ruby_event_store/mappers/type_to_class'
+require 'ruby_event_store/mappers/backward_compatible_builder_with_event_remapping'
 require 'ruby_event_store/mappers/default'
 require 'ruby_event_store/mappers/protobuf'
 require 'ruby_event_store/mappers/null_mapper'
