+++ ruby_event_store/lib/ruby_event_store/mappers/backward_compatible_builder_with_event_remapping.rb
@@ -1,0 +1,21 @@
+module RubyEventStore
+  module Mappers
+    class BackwardCompatibleBuilderWithEventRemapping
+      def initialize(events_class_remapping)
+        unless events_class_remapping.empty?
+          warn <<~EOW
+            events_class_remapping is deprecated. Provide custom event builder instead.
+          EOW
+        end
+        @events_class_remapping = events_class_remapping
+      end
+
+      def call(event_type)
+        type = @events_class_remapping.fetch(event_type) { event_type }
+        ->(args) { Object.const_get(type).new(args) }
+      end
+    end
+
+    private_constant(:BackwardCompatibleBuilderWithEventRemapping)
+  end
+end
